aufile: separate aufile_src.c from aufile.c
