+++ modules/aufile/module.mk
@@ -6,7 +6,7 @@
 
 MOD		:= aufile
 $(MOD)_SRCS	+= aufile.c
-$(MOD)_SRCS	+= aufile_play.c
+$(MOD)_SRCS	+= aufile_play.c aufile_src.c
 $(MOD)_LFLAGS	+=
 
 include mk/mod.mk
